update folder sctructure; fix basic merge errors
