--- src/components/Comment/comment.css
@@ -1,16 +1,0 @@
-.comment {
-    border-bottom: solid 1px #DEE2E7;
-    padding: 2vw 5vw 2vw;
-}
-
-@media screen and (min-width: 481px) {
-    .comment {
-        padding: 2vw 3vw 2vw;
-    }
-}
-
-@media screen and (min-width: 841px) {
-    .comment {
-        padding: 1vw 3vw 1vw;
-    }
-}
