+++ src/main/client/components/Comment/Comment.tsx
@@ -1,20 +1,20 @@
 import * as React from "react";
 
-import { CommentModel } from '../../models/comment';
+import * as Entities from '../../entities';
 
 const styles: any = require('./comment.css');
 const color: any = require('../../theme/color.css');
 
-export class Comment extends React.Component<CommentModel, void> {
+export class Comment extends React.Component<Entities.Comment, void> {
 
     constructor(props) {
         super(props);
     }
 
     public render() {
-        const author = this.props.author || 'KobeJohn',
+        const author = this.props.createdBy || 'KobeJohn',
             content = this.props.content || 'If you come back to this, maybe you should change the accepted answer from diaspora.',
-            date = this.props.date || 'Feb 5 12 at 13:35';
+            date = this.props.createdDate || 'Feb 5 12 at 13:35';
         return (
             <div className={styles.comment}>
                 <span>{content} - </span>
