+++ src/main/client/components/Question/question.scss
