+++ src/main/client/entities/Comment.ts
