+++ src/main/client/entities/Question.ts
@@ -1,0 +1,7 @@
+import * as Entities from '../entities';
+
+export interface Question extends Entities.Post {
+    tags?: string[];
+    title?: string;
+    comments?: Entities.Comment[];
+}
