--- src/main/client/redux/entities/Question.ts
@@ -1,5 +1,0 @@
-import {Post} from '../entities';
-
-export interface Question extends Post {
-    title: string;
-}
