--- src/models/comment.ts
@@ -1,6 +1,0 @@
-export type CommentModel = {
-    id?: number;
-    author?: string;
-    content?: string;
-    date?: string;
-};
