--- src/models/user.ts
@@ -1,6 +1,0 @@
-export type UserModel = {
-    id?: number;
-    author: string;
-    content: string;
-    date: string;
-};
